+++ src/grammar/util/safe_functions.h
