+++ src/grammar/util/vector_functions.c
@@ -2,11 +2,11 @@
 #include <stdio.h>
 #include <limits.h>
 #include <string.h>
-#include "vector_functions.h"
+#include "rolls/vector_functions.h"
 #include "yacc_header.h"
 #include "shared_header.h"
-#include "safe_functions.h"
-#include "rolls/condition_checking.h"
+#include "util/safe_functions.h"
+#include "operations/condition_checking.h"
 
 extern int gnoll_errno;
