+++ src/grammar/util/vector_functions.h
